pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "abc-user-audit"
version = "0.1.0"
authors = [
  { name = "Tyler Norlund", email = "tnorlund@icloud.com" },
]
description = "Data quality and anomaly detection for the ABC user migration take-home"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "pandas>=2.2,<3.0",
    "email-validator>=2.2,<3.0",
    "phonenumberslite>=8.13,<9.0",
    "python-dateutil>=2.9,<3.0",
]
keywords = ["data-quality", "anomalies", "migration", "users", "sweetgreen", "takehome"]

[project.scripts]
abc-audit = "abc_user_audit.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=8.3,<9.0",
    "pytest-cov>=6.0,<7.0",
]
dev = [
    "black>=24.0,<25.0",
    "isort>=5.13,<6.0",
    "mypy>=1.12,<2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/abc_user_audit"]
